# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple medium-like blog API, writen in Go",
        "title": "Go-Blog",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/blog/my": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get information about my blog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get my blog information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Change user's blog name by providing required data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Change user's blog name",
                "parameters": [
                    {
                        "description": "the body to change user's blog name",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateBlogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/blog/{username}": {
            "get": {
                "description": "Get information about user blog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get user blog information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the user",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/blog/{username}/posts/{post_slug}/comments": {
            "get": {
                "description": "Get all comments on a post",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Get comments on a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog owner's username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post slug",
                        "name": "post_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Create a comment on a blog post by providing required data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Create a comment on a blog post",
                "parameters": [
                    {
                        "description": "the body to create a comment",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CommentRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "blog owner's username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post slug",
                        "name": "post_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/blog/{username}/posts/{post_slug}/comments/{comment_id}": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Edit user's comment on a post by providing comment ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Edit user's comment on a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog owner's username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post slug",
                        "name": "post_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body required to modify comment",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CommentRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete comment by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Delete comment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog owner's username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post slug",
                        "name": "post_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
...
        "/my/comments": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get current user's comments on all posts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Get current user's comments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CommentRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateBlogRequest": {
            "type": "object",
            "properties": {
                "new_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "JWT Bearer Token. Need to Login to get the token. Usage: \"Bearer \u003cyour-token-here\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}